Loads of win32 compilation issues fixed.  Most are fairly minor unsigned/signed mismatches because of socket handles, but there were also some order-of-definition issues that were fixed in the FD_SETSIZE definition code.
Fixed a few things that just never worked on windows, like Bu::Thread::yield().

+++ src/stable/archive.cpp
@@ -35,7 +35,7 @@ void Bu::Archive::read( void *pData, size_t nSize )
 	if( nSize == 0 || pData == NULL )
 		return;
 
-	if( rStream.read( (char *)pData, nSize ) < nSize )
+	if( (size_t)rStream.read( (char *)pData, nSize ) < nSize )
 		throw Bu::ExceptionBase("Insufficient data to unarchive object.");
 }
